# fiat-payments-backend/src/api/contracts.js
const express = require('express');
const { ethers } = require('ethers');
const rateLimit = require('express-rate-limit');
const helmet = require('helmet');
const Joi = require('joi');
const winston = require('winston');

const router = express.Router();

// Configuración de seguridad
const limiter = rateLimit({
  windowMs: 15 * 60 * 1000, // 15 minutos
  max: 100, // máximo 100 requests por ventana
  message: {
    error: 'Demasiadas requests desde esta IP',
    retryAfter: '15 minutos'
  },
  standardHeaders: true,
  legacyHeaders: false,
});

// Aplicar rate limiting a todas las rutas
router.use(limiter);
router.use(helmet());

// Logger
const logger = winston.createLogger({
  level: 'info',
  format: winston.format.combine(
    winston.format.timestamp(),
    winston.format.json()
  ),
  transports: [
    new winston.transports.Console(),
    new winston.transports.File({ filename: 'logs/contracts-api.log' })
  ],
});

// Validación de esquemas
const userSchema = Joi.object({
  name: Joi.string().min(2).max(100).required(),
  email: Joi.string().email().required(),
  walletAddress: Joi.string().pattern(/^0x[a-fA-F0-9]{40}$/).required()
});

const courseSchema = Joi.object({
  title: Joi.string().min(5).max(200).required(),
  description: Joi.string().min(10).max(1000).required(),
  price: Joi.number().positive().required(),
  duration: Joi.number().integer().min(1).max(365).required(),
  maxStudents: Joi.number().integer().min(1).max(1000).required(),
  skills: Joi.array().items(Joi.string()).min(1).required(),
  difficulty: Joi.number().integer().min(1).max(5).required()
});

const enrollmentSchema = Joi.object({
  courseId: Joi.number().integer().positive().required(),
  studentAddress: Joi.string().pattern(/^0x[a-fA-F0-9]{40}$/).required()
});

// Middleware de validación
const validateRequest = (schema) => {
  return (req, res, next) => {
    const { error } = schema.validate(req.body);
    if (error) {
      return res.status(400).json({
        error: 'Datos inválidos',
        details: error.details.map(d => d.message)
      });
    }
    next();
  };
};

// Middleware de autenticación
const authenticateUser = async (req, res, next) => {
  try {
    const token = req.headers.authorization?.replace('Bearer ', '');
    if (!token) {
      return res.status(401).json({ error: 'Token de autenticación requerido' });
    }
    
    // Aquí implementarías la verificación del token JWT
    // Por ahora, simulamos la verificación
    req.user = { address: req.body.walletAddress || req.query.walletAddress };
    next();
  } catch (error) {
    logger.error('Error de autenticación:', error);
    res.status(401).json({ error: 'Token inválido' });
  }
};

// ========== ENDPOINTS DE USUARIOS ==========

/**
 * @swagger
 * /api/contracts/users:
 *   post:
 *     summary: Registrar nuevo usuario
 *     tags: [Users]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - name
 *               - email
 *               - walletAddress
 *             properties:
 *               name:
 *                 type: string
 *               email:
 *                 type: string
 *               walletAddress:
 *                 type: string
 */
router.post('/users', validateRequest(userSchema), async (req, res) => {
  try {
    const { name, email, walletAddress } = req.body;
    
    logger.info(`Registrando usuario: ${walletAddress}`);
    
    // Aquí conectarías con el contrato BrainSafes
    // const brainSafes = new ethers.Contract(contractAddress, abi, provider);
    // const tx = await brainSafes.registerUser(name, email, walletAddress);
    
    // Simulación de respuesta
    const userData = {
      id: Date.now(),
      name,
      email,
      walletAddress,
      reputation: 0,
      totalEarned: 0,
      totalSpent: 0,
      joinTimestamp: Date.now(),
      isActive: true,
      achievements: []
    };
    
    res.status(201).json({
      success: true,
      message: 'Usuario registrado exitosamente',
      data: userData
    });
  } catch (error) {
    logger.error('Error registrando usuario:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

/**
 * @swagger
 * /api/contracts/users/{address}:
 *   get:
 *     summary: Obtener perfil de usuario
 *     tags: [Users]
 *     parameters:
 *       - in: path
 *         name: address
 *         required: true
 *         schema:
 *           type: string
 */
router.get('/users/:address', async (req, res) => {
  try {
    const { address } = req.params;
    
    if (!ethers.isAddress(address)) {
      return res.status(400).json({ error: 'Dirección de wallet inválida' });
    }
    
    logger.info(`Obteniendo perfil de usuario: ${address}`);
    
    // Simulación de datos de usuario
    const userProfile = {
      address,
      name: 'Usuario Ejemplo',
      email: 'usuario@ejemplo.com',
      reputation: 850,
      totalEarned: 1500,
      totalSpent: 300,
      joinTimestamp: Date.now() - 86400000 * 30, // 30 días atrás
      isActive: true,
      achievements: [1, 3, 5],
      coursesEnrolled: 5,
      coursesCompleted: 3,
      certificates: 2
    };
    
    res.json({
      success: true,
      data: userProfile
    });
  } catch (error) {
    logger.error('Error obteniendo perfil:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== ENDPOINTS DE CURSOS ==========

/**
 * @swagger
 * /api/contracts/courses:
 *   post:
 *     summary: Crear nuevo curso
 *     tags: [Courses]
 *     security:
 *       - bearerAuth: []
 */
router.post('/courses', authenticateUser, validateRequest(courseSchema), async (req, res) => {
  try {
    const courseData = req.body;
    const instructorAddress = req.user.address;
    
    logger.info(`Creando curso: ${courseData.title} por ${instructorAddress}`);
    
    // Simulación de creación de curso
    const course = {
      id: Date.now(),
      instructor: instructorAddress,
      ...courseData,
      currentStudents: 0,
      totalEarnings: 0,
      isActive: true,
      createdAt: Date.now()
    };
    
    res.status(201).json({
      success: true,
      message: 'Curso creado exitosamente',
      data: course
    });
  } catch (error) {
    logger.error('Error creando curso:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

/**
 * @swagger
 * /api/contracts/courses:
 *   get:
 *     summary: Listar cursos
 *     tags: [Courses]
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *       - in: query
 *         name: instructor
 *         schema:
 *           type: string
 *       - in: query
 *         name: difficulty
 *         schema:
 *           type: integer
 */
router.get('/courses', async (req, res) => {
  try {
    const { page = 1, limit = 10, instructor, difficulty } = req.query;
    
    logger.info(`Listando cursos - Página: ${page}, Límite: ${limit}`);
    
    // Simulación de lista de cursos
    const courses = Array.from({ length: 10 }, (_, i) => ({
      id: i + 1,
      instructor: '0x1234567890123456789012345678901234567890',
      title: `Curso de Blockchain ${i + 1}`,
      description: `Descripción del curso ${i + 1}`,
      price: 100 + (i * 50),
      duration: 30 + (i * 10),
      maxStudents: 100,
      currentStudents: Math.floor(Math.random() * 50),
      totalEarnings: Math.floor(Math.random() * 10000),
      isActive: true,
      skills: ['Solidity', 'Web3', 'DeFi'],
      difficulty: Math.floor(Math.random() * 5) + 1,
      createdAt: Date.now() - (i * 86400000)
    }));
    
    const total = 100; // Total de cursos en la base de datos
    
    res.json({
      success: true,
      data: {
        courses,
        pagination: {
          page: parseInt(page),
          limit: parseInt(limit),
          total,
          pages: Math.ceil(total / limit)
        }
      }
    });
  } catch (error) {
    logger.error('Error listando cursos:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== ENDPOINTS DE INSCRIPCIONES ==========

/**
 * @swagger
 * /api/contracts/enrollments:
 *   post:
 *     summary: Inscribir estudiante en curso
 *     tags: [Enrollments]
 *     security:
 *       - bearerAuth: []
 */
router.post('/enrollments', authenticateUser, validateRequest(enrollmentSchema), async (req, res) => {
  try {
    const { courseId, studentAddress } = req.body;
    
    logger.info(`Inscribiendo estudiante ${studentAddress} en curso ${courseId}`);
    
    // Simulación de inscripción
    const enrollment = {
      courseId: parseInt(courseId),
      student: studentAddress,
      enrolledAt: Date.now(),
      progress: 0,
      score: 0,
      completed: false,
      certificateIssued: false
    };
    
    res.status(201).json({
      success: true,
      message: 'Estudiante inscrito exitosamente',
      data: enrollment
    });
  } catch (error) {
    logger.error('Error inscribiendo estudiante:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== ENDPOINTS DE CERTIFICADOS ==========

/**
 * @swagger
 * /api/contracts/certificates:
 *   post:
 *     summary: Emitir certificado
 *     tags: [Certificates]
 *     security:
 *       - bearerAuth: []
 */
router.post('/certificates', authenticateUser, async (req, res) => {
  try {
    const { courseId, studentAddress, score, metadata } = req.body;
    
    logger.info(`Emitiendo certificado para ${studentAddress} en curso ${courseId}`);
    
    // Simulación de emisión de certificado
    const certificate = {
      id: Date.now(),
      courseId: parseInt(courseId),
      student: studentAddress,
      issuer: req.user.address,
      score: parseInt(score),
      issuedAt: Date.now(),
      metadata: metadata || {},
      tokenId: Math.floor(Math.random() * 1000000)
    };
    
    res.status(201).json({
      success: true,
      message: 'Certificado emitido exitosamente',
      data: certificate
    });
  } catch (error) {
    logger.error('Error emitiendo certificado:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== ENDPOINTS DE GOBIERNO ==========

/**
 * @swagger
 * /api/contracts/governance/proposals:
 *   get:
 *     summary: Listar propuestas de gobierno
 *     tags: [Governance]
 */
router.get('/governance/proposals', async (req, res) => {
  try {
    const { status, page = 1, limit = 10 } = req.query;
    
    logger.info(`Listando propuestas de gobierno - Estado: ${status}`);
    
    // Simulación de propuestas
    const proposals = Array.from({ length: 5 }, (_, i) => ({
      id: i + 1,
      title: `Propuesta ${i + 1}`,
      description: `Descripción de la propuesta ${i + 1}`,
      proposer: '0x1234567890123456789012345678901234567890',
      status: ['active', 'passed', 'failed', 'executed'][i % 4],
      votesFor: Math.floor(Math.random() * 1000),
      votesAgainst: Math.floor(Math.random() * 500),
      createdAt: Date.now() - (i * 86400000 * 7),
      endTime: Date.now() + (i * 86400000 * 3)
    }));
    
    res.json({
      success: true,
      data: proposals
    });
  } catch (error) {
    logger.error('Error listando propuestas:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== ENDPOINTS DE MARKETPLACE ==========

/**
 * @swagger
 * /api/contracts/marketplace/jobs:
 *   get:
 *     summary: Listar trabajos del marketplace
 *     tags: [Marketplace]
 */
router.get('/marketplace/jobs', async (req, res) => {
  try {
    const { category, location, page = 1, limit = 10 } = req.query;
    
    logger.info(`Listando trabajos del marketplace`);
    
    // Simulación de trabajos
    const jobs = Array.from({ length: 10 }, (_, i) => ({
      id: i + 1,
      title: `Trabajo de Blockchain ${i + 1}`,
      description: `Descripción del trabajo ${i + 1}`,
      employer: '0x1234567890123456789012345678901234567890',
      salary: 50000 + (i * 10000),
      location: ['Remoto', 'Nueva York', 'San Francisco', 'Londres'][i % 4],
      category: ['Desarrollo', 'Diseño', 'Marketing', 'Investigación'][i % 4],
      requirements: ['Solidity', 'React', 'Node.js'],
      postedAt: Date.now() - (i * 86400000),
      applications: Math.floor(Math.random() * 50)
    }));
    
    res.json({
      success: true,
      data: jobs
    });
  } catch (error) {
    logger.error('Error listando trabajos:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== ENDPOINTS DE ANALYTICS ==========

/**
 * @swagger
 * /api/contracts/analytics/metrics:
 *   get:
 *     summary: Obtener métricas del sistema
 *     tags: [Analytics]
 */
router.get('/analytics/metrics', async (req, res) => {
  try {
    const { timeframe = '24h' } = req.query;
    
    logger.info(`Obteniendo métricas - Timeframe: ${timeframe}`);
    
    // Simulación de métricas
    const metrics = {
      totalUsers: 15420,
      totalCourses: 342,
      totalCertificates: 8920,
      totalRevenue: 1250000,
      activeEnrollments: 2340,
      completionRate: 0.78,
      averageScore: 4.2,
      platformFee: 31250,
      gasUsed: 4500000,
      transactions: 12500,
      timeframe
    };
    
    res.json({
      success: true,
      data: metrics
    });
  } catch (error) {
    logger.error('Error obteniendo métricas:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== ENDPOINTS DE BRIDGE ==========

/**
 * @swagger
 * /api/contracts/bridge/transfer:
 *   post:
 *     summary: Transferir tokens entre L1 y L2
 *     tags: [Bridge]
 *     security:
 *       - bearerAuth: []
 */
router.post('/bridge/transfer', authenticateUser, async (req, res) => {
  try {
    const { fromChain, toChain, amount, tokenType } = req.body;
    
    logger.info(`Iniciando transferencia bridge: ${fromChain} -> ${toChain}`);
    
    // Simulación de transferencia bridge
    const transfer = {
      id: Date.now(),
      fromChain,
      toChain,
      amount: parseInt(amount),
      tokenType,
      sender: req.user.address,
      status: 'pending',
      createdAt: Date.now(),
      estimatedTime: Date.now() + 300000 // 5 minutos
    };
    
    res.status(201).json({
      success: true,
      message: 'Transferencia bridge iniciada',
      data: transfer
    });
  } catch (error) {
    logger.error('Error en transferencia bridge:', error);
    res.status(500).json({ error: 'Error interno del servidor' });
  }
});

// ========== MANEJO DE ERRORES ==========

// Error handler específico para esta API
router.use((err, req, res, next) => {
  logger.error('Error en API de contratos:', err);
  
  if (err.name === 'ValidationError') {
    return res.status(400).json({
      error: 'Error de validación',
      details: err.message
    });
  }
  
  if (err.name === 'RateLimitExceeded') {
    return res.status(429).json({
      error: 'Límite de requests excedido',
      retryAfter: err.retryAfter
    });
  }
  
  res.status(500).json({
    error: 'Error interno del servidor',
    message: process.env.NODE_ENV === 'development' ? err.message : 'Algo salió mal'
  });
});

module.exports = router;