# fiat-payments-backend/src/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BrainSafes API",
    "description": "API completa para el ecosistema BrainSafes - Educación descentralizada con IA, NFTs y DeFi",
    "version": "1.0.0",
    "contact": {
      "name": "BrainSafes Team",
      "email": "support@brainsafes.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.brainsafes.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "walletAddress": {
            "type": "string"
          },
          "reputation": {
            "type": "number"
          },
          "totalEarned": {
            "type": "number"
          },
          "totalSpent": {
            "type": "number"
          },
          "joinTimestamp": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "instructor": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "maxStudents": {
            "type": "number"
          },
          "currentStudents": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/contracts/users": {
      "post": {
        "summary": "Registrar nuevo usuario",
        "description": "Crear un nuevo usuario en el sistema",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "walletAddress"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "walletAddress": {
                    "type": "string",
                    "pattern": "^0x[a-fA-F0-9]{40}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "429": {
            "description": "Rate limit excedido"
          }
        }
      }
    },
    "/api/contracts/users/{address}": {
      "get": {
        "summary": "Obtener perfil de usuario",
        "description": "Obtener información detallada de un usuario",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Dirección de wallet del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil de usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/contracts/courses": {
      "post": {
        "summary": "Crear nuevo curso",
        "description": "Crear un nuevo curso en el sistema",
        "tags": ["Courses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "price", "duration", "maxStudents", "skills", "difficulty"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 1000
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0
                  },
                  "duration": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 365
                  },
                  "maxStudents": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 1000
                  },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1
                  },
                  "difficulty": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Curso creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Course"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar cursos",
        "description": "Obtener lista paginada de cursos",
        "tags": ["Courses"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Elementos por página"
          },
          {
            "name": "instructor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por instructor"
          },
          {
            "name": "difficulty",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filtrar por dificultad"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cursos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "courses": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Course"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "pages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
} 